upgrade gramchain to v3
