+++ .changeset/afraid-windows-kiss.md
@@ -1,0 +1,5 @@
+---
+'@chainlink/gramchain-adapter': major
+---
+
+Upgraded adapter from v2 to v3 framework
