--- packages/sources/gramchain-test/src/endpoint/index.ts
@@ -1,1 +1,0 @@
-export { endpoint as getgrambalances } from './getgrambalances'
