--- packages/sources/gramchain/schemas/env.json
@@ -1,12 +1,0 @@
-{
-  "$id": "https://external-adapters.chainlinklabs.com/schemas/gramchain-adapter.json",
-  "title": "Chainlink External Adapter for GramChain",
-  "required": [],
-  "type": "object",
-  "properties": {},
-  "allOf": [
-    {
-      "$ref": "https://external-adapters.chainlinklabs.com/schemas/ea-bootstrap.json"
-    }
-  ]
-}
