+++ packages/sources/gramchain/src/endpoint/index.ts
@@ -1,5 +1,1 @@
-import type { TInputParameters as GetGramBalancesInputParameters } from './getgrambalances'
-
-export type TInputParameters = GetGramBalancesInputParameters
-
-export * as getgrambalances from './getgrambalances'
+export { endpoint as getgrambalances } from './getgrambalances'
