+++ packages/sources/gramchain/src/index.ts
@@ -1,9 +1,13 @@
-import { expose } from '@chainlink/ea-bootstrap'
-import { makeExecute, endpointSelector } from './adapter'
-import { makeConfig, NAME } from './config'
-import * as types from './endpoint'
+import { expose, ServerInstance } from '@chainlink/external-adapter-framework'
+import { Adapter } from '@chainlink/external-adapter-framework/adapter'
+import { config } from './config'
+import { getgrambalances } from './endpoint'
 
-const adapterContext = { name: NAME }
+export const adapter = new Adapter({
+  defaultEndpoint: getgrambalances.name,
+  name: 'GRAMCHAIN',
+  config,
+  endpoints: [getgrambalances],
+})
 
-const { server } = expose(adapterContext, makeExecute(), undefined, endpointSelector)
-export { NAME, makeExecute, makeConfig, server, types }
+export const server = (): Promise<ServerInstance | undefined> => expose(adapter)
