+++ packages/sources/gramchain/test/integration/adapter.test.ts
@@ -1,47 +1,45 @@
-import { AdapterRequest } from '@chainlink/ea-bootstrap'
-import process from 'process'
-import { server as startServer } from '../../src'
-import { mockGramChainResponseSuccess } from './fixtures'
-import { DEFAULT_BASE_URL } from '../../src/config'
-import { setupExternalAdapterTest } from '@chainlink/ea-test-helpers'
-import type { SuiteContext } from '@chainlink/ea-test-helpers'
-import { SuperTest, Test } from 'supertest'
+import {
+  TestAdapter,
+  setEnvVariables,
+} from '@chainlink/external-adapter-framework/util/testing-utils'
+import * as nock from 'nock'
+import { mockResponseSuccess } from './fixtures'
 
 describe('execute', () => {
-  const id = '1'
-  const context: SuiteContext = {
-    req: null,
-    server: startServer,
-  }
+  let spy: jest.SpyInstance
+  let testAdapter: TestAdapter
+  let oldEnv: NodeJS.ProcessEnv
 
-  const envVariables = {
-    CACHE_ENABLED: 'false',
-    API_ENDPOINT: process.env.API_ENDPOINT || DEFAULT_BASE_URL,
-    API_VERBOSE: 'true',
-  }
+  beforeAll(async () => {
+    oldEnv = JSON.parse(JSON.stringify(process.env))
+    const mockDate = new Date('2001-01-01T11:11:11.111Z')
+    spy = jest.spyOn(Date, 'now').mockReturnValue(mockDate.getTime())
 
-  setupExternalAdapterTest(envVariables, context)
-  describe('api', () => {
-    const data: AdapterRequest = {
-      id,
-      data: {
+    const adapter = (await import('./../../src')).adapter
+    testAdapter = await TestAdapter.startWithMockedCache(adapter, {
+      testAdapter: {} as TestAdapter<never>,
+    })
+  })
+
+  afterAll(async () => {
+    setEnvVariables(oldEnv)
+    await testAdapter.api.close()
+    nock.restore()
+    nock.cleanAll()
+    spy.mockRestore()
+  })
+
+  describe('getgrambalances endpoint', () => {
+    it('should return success', async () => {
+      const data = {
         custodianID: 'Cache',
         metalCode: 'AU',
         utilizationLockCode: 'Locked',
-      },
-    }
-
-    it('should return success', async () => {
-      mockGramChainResponseSuccess()
-
-      const response = await (context.req as SuperTest<Test>)
-        .post('/')
-        .send(data)
-        .set('Accept', '*/*')
-        .set('Content-Type', 'application/json')
-        .expect('Content-Type', /json/)
-        .expect(200)
-      expect(response.body).toMatchSnapshot()
+      }
+      mockResponseSuccess()
+      const response = await testAdapter.request(data)
+      expect(response.statusCode).toBe(200)
+      expect(response.json()).toMatchSnapshot()
     })
   })
 })
